requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
crispy-bootstrap5==2025.6
Django==5.2.4
django-crispy-forms==2.4
django-jazzmin==3.0.1
django-recaptcha==4.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
qrcode==8.2
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
